+++ scripts/__init__.py
@@ -6,6 +6,7 @@ from .wechat_client import WeChatClient, WeChatAPIError, create_client, load_dot
 from .material_manager import MaterialManager, create_material_manager
 from .draft_manager import DraftManager, create_draft_manager, create_simple_article
 from .stats_manager import StatsManager, create_stats_manager
+from .html_submitter import submit_html_draft, HtmlSubmitError, ImageUploadError
 
 __all__ = [
     "WeChatClient",
@@ -19,4 +20,7 @@ __all__ = [
     "create_simple_article",
     "StatsManager",
     "create_stats_manager",
+    "submit_html_draft",
+    "HtmlSubmitError",
+    "ImageUploadError",
 ]